+++ src/gql/arguments/TopPagesArguments.php
@@ -2,6 +2,7 @@
 
 namespace vnali\counter\gql\arguments;
 
+use GraphQL\Type\Definition\InputObjectType;
 use GraphQL\Type\Definition\Type;
 
 class TopPagesArguments
@@ -15,6 +16,16 @@ class TopPagesArguments
                 'name' => 'dateRange',
                 'type' => Type::nonNull(Type::STRING()),
                 'description' => 'Date Range',
+            ],
+            'showElementTitle' => [
+                'name' => 'showElementTitle',
+                'type' => Type::BOOLEAN(),
+                'description' => 'Show element title',
+            ],
+            'filters' => [
+                'name' => 'filters',
+                'type' => self::getFiltersInputType(),
+                'description' => 'Filter elements including section handles',
             ],
             'siteId' => [
                 'name' => 'siteId',
@@ -32,5 +43,22 @@ class TopPagesArguments
                 'description' => 'Random unique value to prevent reading result from cache',
             ],
         ];
+    }
+
+    private static function getFiltersInputType(): InputObjectType
+    {
+        return new InputObjectType([
+            'name' => 'FiltersInput',
+            'fields' => [
+                'sectionHandles' => [
+                    'type' => Type::listOf(Type::STRING()),
+                    'description' => 'Array of section handles',
+                ],
+                'items' => [
+                    'type' => Type::listOf(Type::STRING()),
+                    'description' => 'Array of items: page, entry, category',
+                ],
+            ],
+        ]);
     }
 }