+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 ## Unreleased
 
 - Added a utility to import statistics from the Views Work plugin. [(#1)](https://github.com/vnali/craft-counter-docs/issues/1)
-- Added counter statistics to the entry and category index pages. [(#1)](https://github.com/vnali/craft-counter-docs/issues/1)
+- Added counter statistics to the entry and category index pages.
 - Added counter statistics to the entry and category GQL queries.
+- Added the possibility for top pages to return the element title instead of the page URL for pages that return an element.
+- Changed the top page statistics to filter pages related to the elements.
 - Fixed a bug where widgets are not loaded when the plugin settings are not configured.
-- Change the label of column in the page visits widget from value to Visits.
+- Minimally change the label of column in the page visits widget from value to Visits.
 
 ## 1.0.0-beta.2 - 2025-01-11